pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-windows-admin"
version = "0.1.0"
description = "MCP server for Windows administration: Antivirus/GameBooster profiles, diagnostics, and system insights"
authors = [
  { name = "Community" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "mcp[cli]>=1.0.0",
  "psutil>=5.9.5",
  "pywin32>=306; platform_system=='Windows'",
  "wmi>=1.5.1; platform_system=='Windows'",
  "pydantic>=2.7.0",
  "typer>=0.12.0",
  "httpx>=0.27.0",
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "aiofiles>=23.2.1",
]

[project.optional-dependencies]
 dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.5.0",
  "mypy>=1.8.0",
  "fastmcp>=2.0.0",
 ]

 yara = [
  "yara-python>=4.5.0",
 ]

 adk = [
  "google-adk>=0.3.0",
 ]

[project.urls]
Homepage = "https://github.com/your-org/mcp-windows-admin"
Issues = "https://github.com/your-org/mcp-windows-admin/issues"

[tool.setuptools]
packages = ["mcp_win_admin", "dashboard_api"]
include-package-data = true

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.10"
strict = false
warn_unused_ignores = true
